refactor: move gh-actions artifact/cache fs writes to core
